Addressed feedbacks: Moved the interop depenedency helper functions to InteropStubManager
